// types/supabase.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  graphql_public: {
    Tables: {
      [_ in never]: never
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      graphql: {
        Args: {
          operationName?: string
          query?: string
          variables?: Json
          extensions?: Json
        }
        Returns: Json
      }
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
  public: {
    Tables: {
      product_search: {
        Row: {
          asin: string | null
          created_at: string
          id: number
          search_id: number | null
        }
        Insert: {
          asin?: string | null
          created_at?: string
          id?: number
          search_id?: number | null
        }
        Update: {
          asin?: string | null
          created_at?: string
          id?: number
          search_id?: number | null
        }
        Relationships: [
          {
            foreignKeyName: "product_search_asin_fkey"
            columns: ["asin"]
            isOneToOne: false
            referencedRelation: "products"
            referencedColumns: ["asin"]
          },
          {
            foreignKeyName: "product_search_search_id_fkey"
            columns: ["search_id"]
            isOneToOne: false
            referencedRelation: "searches"
            referencedColumns: ["id"]
          },
        ]
      }
      products: {
        Row: {
          asin: string
          created_at: string
          currency: string | null
          final_price: number
          image: string | null
          name: string | null
          updated_at: string
          url: string | null
        }
        Insert: {
          asin: string
          created_at?: string
          currency?: string | null
          final_price?: number
          image?: string | null
          name?: string | null
          updated_at?: string
          url?: string | null
        }
        Update: {
          asin?: string
          created_at?: string
          currency?: string | null
          final_price?: number
          image?: string | null
          name?: string | null
          updated_at?: string
          url?: string | null
        }
        Relationships: []
      }
      searches: {
        Row: {
          created_at: string
          id: number
          is_tracked: boolean
          last_scraped_at: string | null
          query: string
          snapshot_id: string | null
          status: string
          user_id: string | null
        }
        Insert: {
          created_at?: string
          id?: number
          is_tracked?: boolean
          last_scraped_at?: string | null
          query: string
          snapshot_id?: string | null
          status?: string
          user_id?: string | null
        }
        Update: {
          created_at?: string
          id?: number
          is_tracked?: boolean
          last_scraped_at?: string | null
          query?: string
          snapshot_id?: string | null
          status?: string
          user_id?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "searches_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          },
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type PublicSchema = Database[Extract<keyof Database, "public">]

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
        PublicSchema["Views"])
    ? (PublicSchema["Tables"] &
        PublicSchema["Views"])[PublicTableNameOrOptions] extends {
        Row: infer R
      }
      ? R
      : never
    : never
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
    ? PublicSchema["Enums"][PublicEnumNameOrOptions]
    : never